Functions use two parameters to calculate power and product of array
Fourth solution for eighth exercise involved writing a function that takes an array parameter and multiplies all of the numbers together.
Fifth solution for eighth exercise involved writing a function that takes two parameters and, using the first as a base and the second as an exponential, returns the power.